+++ src/minecraft/ip/industrialProcessing/power/plants/IHeatable.java
@@ -1,5 +1,5 @@
 package ip.industrialProcessing.power.plants;
 
 public interface IHeatable {
-	void addHeat(int heat);
+	void addHeat(float heat);
 }
